+++ lib/dbee/model.rb
@@ -8,6 +8,7 @@
 #
 
 require_relative 'model/constraints'
+require_relative 'model/partitioner'
 
 module Dbee
   # In DB terms, a Model is usually a table, but it does not have to be.  You can also re-model
@@ -16,53 +17,48 @@ module Dbee
     extend Forwardable
     acts_as_hashable
 
-    JOIN_CHAR = '.'
-
-    private_constant :JOIN_CHAR
-
     class ModelNotFoundError < StandardError; end
 
-    attr_reader :constraints, :name, :table
+    attr_reader :constraints, :filters, :name, :partitioners, :table
 
-    def_delegator :models_by_name, :values, :models
+    def_delegator :models_by_name,  :values,  :models
+    def_delegator :models,          :sort,    :sorted_models
+    def_delegator :constraints,     :sort,    :sorted_constraints
+    def_delegator :partitioners,    :sort,    :sorted_partitioners
 
-    def_delegator :models, :sort, :sorted_models
-
-    def_delegator :constraints, :sort, :sorted_constraints
-
-    def initialize(name:, constraints: [], models: [], table: '')
+    def initialize(name:, constraints: [], models: [], partitioners: [], table: '')
       raise ArgumentError, 'name is required' if name.to_s.empty?
 
       @name           = name.to_s
-      @constraints    = Constraints.array(constraints)
+      @constraints    = Constraints.array(constraints).uniq
       @models_by_name = name_hash(Model.array(models))
+      @partitioners   = Partitioner.array(partitioners).uniq
       @table          = table.to_s.empty? ? @name : table.to_s
 
       freeze
     end
 
-    def name_hash(array)
-      array.map { |a| [a.name, a] }.to_h
-    end
-
-    def ancestors(parts = [], alias_chain = [], found = {})
+    # This recursive method will walk a path of model names (parts) and return back a
+    # flattened hash instead of a nested object structure.
+    # The hash key will be an array of strings (model names) and the value will be the
+    # identified model.
+    def ancestors!(parts = [], visited_parts = [], found = {})
       return found if Array(parts).empty?
 
-      alias_chain = [] if Array(alias_chain).empty?
+      # Take the first entry in parts
+      model_name = parts.first.to_s
 
-      model_name = parts.first
+      # Ensure we have it registered as a child, or raise error
+      model = assert_model(model_name, visited_parts)
 
-      model = models_by_name[model_name.to_s]
+      # Push onto visited list
+      visited_parts += [model_name]
 
-      raise ModelNotFoundError, "Cannot traverse: #{model_name}" unless model
+      # Add found model to flattened structure
+      found[visited_parts] = model
 
-      new_alias_chain = alias_chain + [model_name]
-
-      new_alias = new_alias_chain.join(JOIN_CHAR)
-
-      found[new_alias] = model
-
-      model.ancestors(parts[1..-1], new_alias_chain, found)
+      # Recursively call for next parts in the chain
+      model.ancestors!(parts[1..-1], visited_parts, found)
     end
 
     def ==(other)
@@ -70,6 +66,7 @@ module Dbee
         other.name == name &&
         other.table == table &&
         other.sorted_constraints == sorted_constraints &&
+        other.sorted_partitioners == sorted_partitioners &&
         other.sorted_models == sorted_models
     end
     alias eql? ==
@@ -81,5 +78,14 @@ module Dbee
     private
 
     attr_reader :models_by_name
+
+    def assert_model(model_name, visited_parts)
+      models_by_name[model_name] ||
+        raise(ModelNotFoundError, "Missing: #{model_name}, after: #{visited_parts}")
+    end
+
+    def name_hash(array)
+      array.map { |a| [a.name, a] }.to_h
+    end
   end
 end